Move urls out of constants.ts, move constants to folder
