+++ src/backend/__tests__/main_window.test.ts
@@ -1,7 +1,7 @@
 import { createMainWindow, sendFrontendMessage } from '../main_window'
 import { BrowserWindow, Display, screen } from 'electron'
 import { overrideProcessPlatform } from './constants.test'
-import { configStore } from 'backend/key_value_stores'
+import { configStore } from 'backend/constants/key_value_stores'
 
 jest.mock('../logger/logfile')
