+++ src/backend/constants/key_value_stores.ts
@@ -1,4 +1,4 @@
-import { TypeCheckedStoreBackend } from './electron_store'
+import { TypeCheckedStoreBackend } from '../electron_store'
 
 export const configStore = new TypeCheckedStoreBackend('configStore', {
   cwd: 'store'
