+++ src/backend/constants/urls.ts
@@ -1,0 +1,19 @@
+export const epicLoginUrl = 'https://legendary.gl/epiclogin'
+export const sidInfoUrl =
+  'https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/wiki/How-To:-Epic-Alternative-Login'
+export const heroicGithubURL =
+  'https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/releases/latest'
+export const GITHUB_API =
+  'https://api.github.com/repos/Heroic-Games-Launcher/HeroicGamesLauncher/releases'
+export const supportURL =
+  'https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/blob/main/Support.md'
+export const discordLink = 'https://discord.gg/rHJ2uqdquK'
+export const wikiLink =
+  'https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/wiki'
+export const weblateUrl =
+  'https://hosted.weblate.org/projects/heroic-games-launcher'
+export const kofiPage = 'https://ko-fi.com/heroicgames'
+export const patreonPage = 'https://www.patreon.com/heroicgameslauncher'
+export const wineprefixFAQ = 'https://wiki.winehq.org/FAQ#Wineprefixes'
+export const customThemesWikiLink =
+  'https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/wiki/Custom-Themes'
