+++ src/backend/launcher.ts
@@ -95,7 +95,7 @@ import { app, powerSaveBlocker } from 'electron'
 import gogPresence from './storeManagers/gog/presence'
 import { updateGOGPlaytime } from './storeManagers/gog/games'
 import { addRecentGame } from './recent_games/recent_games'
-import { tsStore } from './key_value_stores'
+import { tsStore } from './constants/key_value_stores'
 
 let powerDisplayId: number | null
