+++ src/backend/recent_games/recent_games.ts
@@ -2,7 +2,7 @@ import { GameInfo, RecentGame } from 'common/types'
 import { backendEvents } from '../backend_events'
 import { sendFrontendMessage } from '../main_window'
 import { GlobalConfig } from '../config'
-import { configStore } from 'backend/key_value_stores'
+import { configStore } from 'backend/constants/key_value_stores'
 
 const maxRecentGames = async () => {
   const { maxRecentGames } = GlobalConfig.get().getSettings()
