+++ src/backend/shortcuts/nonesteamgame/nonesteamgame.ts
@@ -24,7 +24,7 @@ import i18next from 'i18next'
 import { notify, showDialogBoxModalAuto } from '../../dialog/dialog'
 import { GlobalConfig } from '../../config'
 import { getWikiGameInfo } from 'backend/wiki_game_info/wiki_game_info'
-import { tsStore } from 'backend/key_value_stores'
+import { tsStore } from 'backend/constants/key_value_stores'
 
 const getSteamUserdataDir = async () => {
   const { defaultSteamPath } = GlobalConfig.get().getSettings()
