+++ src/backend/storeManagers/legendary/games.ts
@@ -90,7 +90,7 @@ import { getUmuId } from 'backend/wiki_game_info/umu/utils'
 import thirdParty from './thirdParty'
 import { Path } from 'backend/schemas'
 import { mkdirSync } from 'fs'
-import { configStore } from 'backend/key_value_stores'
+import { configStore } from 'backend/constants/key_value_stores'
 
 /**
  * Alias for `LegendaryLibrary.listUpdateableGames`
