+++ src/backend/tray_icon/__tests__/tray_icon.test.ts
@@ -4,7 +4,7 @@ import { backendEvents } from '../../backend_events'
 import { GlobalConfig } from '../../config'
 import { RecentGame } from 'common/types'
 import i18next from 'i18next'
-import { configStore } from 'backend/key_value_stores'
+import { configStore } from 'backend/constants/key_value_stores'
 
 jest.mock('../../logger/logfile')
 jest.mock('../../config')
